feat: Cleanup database schema, use Long instead of Double, LocalDate instead of Instant and remove UserId from PensionStatement

+++ backend/src/main/java/com/pensionplanner/audit/OtherIncomeAudit.java
@@ -34,7 +34,7 @@ public class OtherIncomeAudit {
     private String name;
 
     @Column(name = "annualAmount")
-    private Double annualAmount;
+    private Long annualAmount;
 
     @Column(name = "notes")
     private String notes;
@@ -87,11 +87,11 @@ public class OtherIncomeAudit {
         this.name = name;
     }
 
-    public Double getAnnualAmount() {
+    public Long getAnnualAmount() {
         return annualAmount;
     }
 
-    public void setAnnualAmount(Double annualAmount) {
+    public void setAnnualAmount(Long annualAmount) {
         this.annualAmount = annualAmount;
     }
